+++ wx-server/src/main/java/com/wx/dao/WxResourceDao.java
@@ -3,6 +3,6 @@ package com.wx.dao;
 import com.wx.entity.WxResource;
 import org.springframework.data.repository.CrudRepository;
 
-public interface WxResourceDao extends CrudRepository<WxResource,Integer> {
-    WxResource findById(int id);
+public interface WxResourceDao extends CrudRepository<WxResource,Long> {
+    WxResource findById(Long id);
 }
